move classes from core.py to base.py: Index, VectorData, VectorIndex, ElementIdentifiers, DynamicTable, DynamicTable
